feat: move logplex/encoding and add license
